rework notebook tests
