+++ e2e/app/page/notebook-preview-page.ts
@@ -42,6 +42,7 @@ export default class NotebookPreviewPage extends AuthenticatedPage {
 
     const notebookPage = new NotebookPage(this.page, notebookName);
     await notebookPage.waitForLoad();
+    await notebookPage.waitForKernelIdle();
     return notebookPage;
   }
