+++ e2e/resources/python-code/import-os.py
@@ -1,5 +1,7 @@
 import sys
 import os
+
+print(sys.version)
 print(sys.executable)
 print(os.getenv('OWNER_EMAIL'))
 print(os.getenv('WORKSPACE_CDR'))
@@ -7,3 +9,5 @@ print(os.getenv('WORKSPACE_NAMESPACE'))
 print(os.getenv('GOOGLE_PROJECT')) # same as WORKSPACE_NAMESPACE
 print(os.getenv('CLUSTER_NAME'))
 print(os.getenv('WORKSPACE_BUCKET'))
+
+print('success')
