+++ e2e/resources/r-code/calculate-max.R
@@ -1,0 +1,7 @@
+sessionInfo()
+
+name <- c("Jon", "Bill", "Maria")
+age <- c(1, 10, 20)
+data <- data.frame(name, age)
+
+print(max(data$age),quote=FALSE)
