+++ e2e/resources/r-code/import-libs.R
@@ -30,3 +30,5 @@ library(reticulate)
 library(scales)
 
 pd <- reticulate::import("pandas")
+
+print('success',quote=FALSE)
