+++ e2e/resources/r-code/sys-print.R
@@ -4,3 +4,4 @@ print(Sys.getenv('WORKSPACE_NAMESPACE'))
 print(Sys.getenv('GOOGLE_PROJECT')) # same as WORKSPACE_NAMESPACE
 print(Sys.getenv('CLUSTER_NAME'))
 print(Sys.getenv('WORKSPACE_BUCKET'))
+print('success',quote=FALSE)
